Backend/requirements.txt
flask
requests
flask-swagger-ui
pandas
scikit-learn
flask-cors
python-dotenv